refactored the TextArea component, ColorPalette
